Set debug value to include error messages of routes cause an error

To hide any error messages in production, set the debug value to FALSE. The debug value is enabled by default for interactive() sessions.

pr_set_debug(pr, debug = interactive())

Arguments

pr

A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.

debug

TRUE provides more insight into your API errors.

Value

The Plumber router with the new debug setting.

Examples

if (FALSE) {
# Will contain the original error message
pr() %>%
  pr_set_debug(TRUE) %>%
  pr_get("/boom", function() stop("boom")) %>%
  pr_run()

# Will NOT contain an error message
pr() %>%
  pr_set_debug(FALSE) %>%
  pr_get("/boom", function() stop("boom")) %>%
  pr_run()
}