Static file router

Static file router

Static file router

Details

Creates a router that is backed by a directory of files on disk.

Super classes

plumber::hookable -> plumber::plumber -> plumberstatic

Methods

Public methods

PlumberStatic$new()

PlumberStatic$print()

PlumberStatic$clone()

Inherited methods

plumber::hookable$registerHooks()

plumber::plumber$addAssets()

plumber::plumber$addEndpoint()

plumber::plumber$addFilter()

plumber::plumber$addGlobalProcessor()

plumber::plumber$call()

plumber::plumber$filter()

plumber::plumber$get_api_spec()

plumber::plumber$get_debug()

plumber::plumber$handle()

plumber::plumber$mount()

plumber::plumber$onHeaders()

plumber::plumber$onWSOpen()

plumber::plumber$openAPIFile()

plumber::plumber$registerHook()

plumber::plumber$remove_handle()

plumber::plumber$route()

plumber::plumber$run()

plumber::plumber$serve()

plumber::plumber$set404Handler()

plumber::plumber$setErrorHandler()

plumber::plumber$setSerializer()

plumber::plumber$set_api_spec()

plumber::plumber$set_debug()

plumber::plumber$set_parsers()

plumber::plumber$set_ui()

plumber::plumber$set_ui_callback()

plumber::plumber$swaggerFile()

plumber::plumber$unmount()

Method new()

Create a new plumberstatic router

Usage

PlumberStatic$new(direc, options)

Arguments

direc

a path to an asset directory.

options

options to be evaluated in the plumberstatic router environment

Returns

A new plumberstatic router

Method print()

Print representation of PlumberStatic() router.

Usage

PlumberStatic$print(prefix = "", topLevel = TRUE, ...)

Arguments

prefix

a character string. Prefix to append to representation.

topLevel

a logical value. When method executed on top level router, set to TRUE.

...

additional arguments for recursive calls

Returns

A terminal friendly represention of a PlumberStatic() router.

Method clone()

The objects of this class are cloneable with this method.

Usage

PlumberStatic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.