Static file router

Static file router

Static file router

Details

Creates a router that is backed by a directory of files on disk.

Super classes

plumber::Hookable -> plumber::Plumber -> PlumberStatic

Methods

Public methods

PlumberStatic$new()

PlumberStatic$print()

PlumberStatic$clone()

Inherited methods

plumber::Hookable$registerHooks()

plumber::Plumber$addAssets()

plumber::Plumber$addEndpoint()

plumber::Plumber$addFilter()

plumber::Plumber$addGlobalProcessor()

plumber::Plumber$call()

plumber::Plumber$filter()

plumber::Plumber$getApiSpec()

plumber::Plumber$getDebug()

plumber::Plumber$handle()

plumber::Plumber$mount()

plumber::Plumber$onHeaders()

plumber::Plumber$onWSOpen()

plumber::Plumber$openAPIFile()

plumber::Plumber$registerHook()

plumber::Plumber$removeHandle()

plumber::Plumber$route()

plumber::Plumber$run()

plumber::Plumber$serve()

plumber::Plumber$set404Handler()

plumber::Plumber$setApiSpec()

plumber::Plumber$setDebug()

plumber::Plumber$setDocs()

plumber::Plumber$setDocsCallback()

plumber::Plumber$setErrorHandler()

plumber::Plumber$setParsers()

plumber::Plumber$setSerializer()

plumber::Plumber$swaggerFile()

plumber::Plumber$unmount()

Method new()

Create a new PlumberStatic router

Usage

PlumberStatic$new(direc, options)

Arguments

direc

a path to an asset directory.

options

options to be evaluated in the PlumberStatic router environment

Returns

A new PlumberStatic router

Method print()

Print representation of PlumberStatic() router.

Usage

PlumberStatic$print(prefix = "", topLevel = TRUE, ...)

Arguments

prefix

a character string. Prefix to append to representation.

topLevel

a logical value. When method executed on top level router, set to TRUE.

...

additional arguments for recursive calls

Returns

A terminal friendly representation of a PlumberStatic() router.

Method clone()

The objects of this class are cloneable with this method.

Usage

PlumberStatic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.