Deprecated R6 functions

Deprecated R6 functions

Deprecated R6 functions

Deprecated R6 functions

Deprecated R6 functions

Functions

hookable: See Hookable()

plumber: See Plumber()

Super class

plumber::Hookable -> hookable

Methods

Public methods

hookable$new()

hookable$clone()

Inherited methods

plumber::Hookable$registerHook()

plumber::Hookable$registerHooks()

Method new()

Initialize a new hookable. Throws deprecated warning prompting user to use Hookable

Usage

hookable$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

hookable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

plumber::Hookable -> plumber::Plumber -> plumber

Methods

Public methods

plumber$new()

plumber$clone()

Inherited methods

plumber::Hookable$registerHooks()

plumber::Plumber$addAssets()

plumber::Plumber$addEndpoint()

plumber::Plumber$addFilter()

plumber::Plumber$addGlobalProcessor()

plumber::Plumber$call()

plumber::Plumber$filter()

plumber::Plumber$getApiSpec()

plumber::Plumber$getDebug()

plumber::Plumber$handle()

plumber::Plumber$mount()

plumber::Plumber$onHeaders()

plumber::Plumber$onWSOpen()

plumber::Plumber$openAPIFile()

plumber::Plumber$print()

plumber::Plumber$registerHook()

plumber::Plumber$removeHandle()

plumber::Plumber$route()

plumber::Plumber$run()

plumber::Plumber$serve()

plumber::Plumber$set404Handler()

plumber::Plumber$setApiSpec()

plumber::Plumber$setDebug()

plumber::Plumber$setDocs()

plumber::Plumber$setDocsCallback()

plumber::Plumber$setErrorHandler()

plumber::Plumber$setParsers()

plumber::Plumber$setSerializer()

plumber::Plumber$swaggerFile()

plumber::Plumber$unmount()

Method new()

Initialize a new plumber. Throws deprecated warning prompting user to use Plumber

Usage

plumber$new(...)

Arguments

...

params passed in to Plumber$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

plumber$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.