Plumber Serializers

Serializers are used in Plumber to transform the R object produced by a filter/endpoint into an HTTP response that can be returned to the client. See here for more details on Plumber serializers and how to customize their behavior.

serializer_headers(headers = list(), serialize_fn = identity)

serializer_content_type(type, serialize_fn = identity)

serializer_csv(...)

serializer_html()

serializer_json(...)

serializer_unboxed_json(auto_unbox = TRUE, ...)

serializer_rds(version = "2", ascii = FALSE, ...)

serializer_feather()

serializer_yaml(...)

serializer_text(..., serialize_fn = as.character)

serializer_format(...)

serializer_print(...)

serializer_cat(...)

serializer_htmlwidget(...)

Arguments

headers
list() of headers to add to the response object

serialize_fn
Function to serialize the data. The result object will be converted to a character string. Ex: jsonlite::toJSON().

type
The value to provide for the Content-Type HTTP header.

...
extra arguments supplied to respective internal serialization function.

auto_unbox
automatically unbox all atomic vectors of length 1. It is usually safer to avoid this and instead use the unbox function to unbox individual elements. An exception is that objects of class AsIs (i.e. wrapped in I()) are not automatically unboxed. This is a way to mark single values as length-1 arrays.

version
the workspace format version to use. NULL specifies the current default version (3). The only other supported value is 2, the default from R 1.4.0 to R 3.5.0.

ascii
a logical. If TRUE or NA, an ASCII representation is written; otherwise (default) a binary one. See also the comments in the help for save.

Functions

serializer_headers: Add a static list of headers to each return value. Will add Content-Disposition header if a value is the result of as_attachment().

serializer_content_type: Adds a Content-Type header to the response object

serializer_csv: CSV serializer. See readr::format_csv() for more details.

serializer_html: HTML serializer

serializer_json: JSON serializer. See jsonlite::toJSON() for more details.

serializer_unboxed_json: JSON serializer with auto_unbox defaulting to TRUE. See jsonlite::toJSON() for more details.

serializer_rds: RDS serializer. See serialize() for more details.

serializer_feather: feather serializer. See feather::write_feather for more details.

serializer_yaml: YAML serializer. See yaml::as.yaml() for more details.

serializer_text: Text serializer. See as.character() for more details.

serializer_format: Text serializer. See format() for more details.

serializer_print: Text serializer. Captures the output of print()

serializer_cat: Text serializer. Captures the output of cat()

serializer_htmlwidget: htmlwidget serializer. See htmlwidgets::saveWidget() for more details.